+++ src/test/java/localgaji/albbaim/auth/AuthServiceTest.java
@@ -15,8 +15,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static localgaji.albbaim.utils.Samples.someKakaoAuth;
-import static localgaji.albbaim.utils.Samples.someUser;
+import static localgaji.albbaim.__utils__.Samples.someKakaoAuth;
+import static localgaji.albbaim.__utils__.Samples.someUser;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
