+++ src/test/java/localgaji/albbaim/user/UserControllerTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import static localgaji.albbaim.utils.Samples.someUser;
+import static localgaji.albbaim.__utils__.Samples.someUser;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
